+++ src/components/month/calendarMonthView.component.ts
@@ -46,7 +46,13 @@ import { CalendarEventTimesChangedEvent } from '../../interfaces/calendarEventTi
   template: `
     <div class="cal-month-view">
       <div class="cal-cell-row cal-header">
-        <div class="cal-cell" *ngFor="let header of columnHeaders">
+        <div
+          class="cal-cell"
+          *ngFor="let header of columnHeaders"
+          [class.cal-past]="header.isPast"
+          [class.cal-today]="header.isToday"
+          [class.cal-future]="header.isFuture"
+          [class.cal-weekend]="header.isWeekend">
           {{ header.date | calendarDate:'monthViewColumnHeader':locale }}
         </div>
       </div>