+++ test/calendarDayView.component.spec.ts
@@ -6,7 +6,8 @@ import {
 } from '@angular/core/testing';
 import * as moment from 'moment';
 import { expect } from 'chai';
-import { DraggableHelper } from 'angular-draggable-droppable';
+import * as sinon from 'sinon';
+import { DraggableHelper, DragAndDropModule } from 'angular-draggable-droppable';
 import {
   CalendarEventTitleFormatter,
   CalendarEvent,
@@ -19,12 +20,20 @@ import {
 import { CalendarDayViewComponent } from './../src/components/day/calendarDayView.component';
 import { Subject } from 'rxjs/Rx';
 import { spy } from 'sinon';
-import { triggerDomEvent } from './util';
+import { triggerDomEvent, ExternalEventComponent } from './util';
 
 describe('CalendarDayViewComponent component', () => {
 
   beforeEach(() => {
-    TestBed.configureTestingModule({imports: [CalendarModule]});
+    TestBed.configureTestingModule({
+      imports: [
+        CalendarModule,
+        DragAndDropModule
+      ],
+      declarations: [
+        ExternalEventComponent
+      ]
+    });
     TestBed.configureCompiler({
       providers: [
         DraggableHelper,
@@ -478,6 +487,42 @@ describe('CalendarDayViewComponent component', () => {
     triggerDomEvent('mouseup', document.body, {clientY: rect.top - 60, clientX: rect.left + 10});
     fixture.detectChanges();
     fixture.destroy();
+  });
+
+  it('should allow external events to be dropped on the day view', () => {
+    const fixture: ComponentFixture<CalendarDayViewComponent> = TestBed.createComponent(CalendarDayViewComponent);
+    fixture.componentInstance.viewDate = new Date('2016-06-27');
+    fixture.componentInstance.events = [];
+    fixture.componentInstance.ngOnChanges({viewDate: {}, events: {}});
+    fixture.detectChanges();
+    document.body.appendChild(fixture.nativeElement);
+
+    const externalEventFixture: ComponentFixture<ExternalEventComponent> = TestBed.createComponent(ExternalEventComponent);
+    externalEventFixture.detectChanges();
+    document.body.appendChild(externalEventFixture.nativeElement);
+
+    const event: HTMLElement = externalEventFixture.nativeElement.querySelector('.external-event');
+    const eventPosition: ClientRect = event.getBoundingClientRect();
+
+    const segments: any[] = Array.from(fixture.nativeElement.querySelectorAll('.cal-hour-segment'));
+    const segment: HTMLElement = segments[2].parentNode;
+    const segmentPosition: ClientRect = segment.getBoundingClientRect();
+
+    const eventDropped: sinon.SinonSpy = sinon.spy();
+    fixture.componentInstance.eventTimesChanged.subscribe(eventDropped);
+    triggerDomEvent('mousedown', event, {clientY: eventPosition.top, clientX: eventPosition.left});
+    fixture.detectChanges();
+    triggerDomEvent('mousemove', document.body, {clientY: segmentPosition.top, clientX: segmentPosition.left});
+    fixture.detectChanges();
+    expect(segment.classList.contains('cal-drag-over')).to.be.true;
+    triggerDomEvent('mouseup', document.body, {clientY: segmentPosition.top, clientX: segmentPosition.left});
+    fixture.detectChanges();
+    fixture.destroy();
+    externalEventFixture.destroy();
+    expect(eventDropped).to.have.been.calledWith({
+      event: externalEventFixture.componentInstance.event,
+      newStart: moment('2016-06-27').startOf('day').add(1, 'hours').toDate()
+    });
   });
 
 });